Почти реализовано API TEST. Отказываются добавляться профессии

+++ questionnaire-be/src/main/java/ru/obrubov/questionnaire/data/access/TestDataAccess.java
@@ -20,12 +20,13 @@ public class TestDataAccess {
     }
 
     public Test getOne() {
-        //TODO
-        return null;
+        return testRepository.getOne();
     }
 
     public void delete() {
         Test testToDelete = getOne();
-        testRepository.delete(testToDelete);
+        if(testToDelete != null) {
+            testRepository.delete(testToDelete);
+        }
     }
 }
